+++ src/bl64-pkg.bash
@@ -1,7 +1,7 @@
 #######################################
 # BashLib64 / Module / Functions / Manage native OS packages
 #
-# Version: 1.10.0
+# Version: 1.12.0
 #######################################
 
 #######################################
@@ -20,6 +20,7 @@
 #######################################
 function bl64_pkg_deploy() {
   bl64_dbg_lib_show_function "$@"
+
   bl64_pkg_prepare &&
     bl64_pkg_install "$@" &&
     bl64_pkg_cleanup
@@ -40,34 +41,24 @@ function bl64_pkg_deploy() {
 #######################################
 function bl64_pkg_prepare() {
   bl64_dbg_lib_show_function
-  local verbose=''
-
-  bl64_check_privilege_root || return $?
-
-  if bl64_dbg_lib_command_enabled; then
-    verbose="$BL64_PKG_SET_VERBOSE"
-  else
-    verbose="$BL64_PKG_SET_QUIET"
-  fi
 
   bl64_msg_show_task "$_BL64_PKG_TXT_PREPARE"
   # shellcheck disable=SC2086
   case "$BL64_OS_DISTRO" in
   ${BL64_OS_FD}-* | ${BL64_OS_RHEL}-* | ${BL64_OS_ALM}-* | ${BL64_OS_RCK}-* | ${BL64_OS_CNT}-8.* | ${BL64_OS_CNT}-9.* | ${BL64_OS_OL}-8.*)
-    "$BL64_PKG_CMD_DNF" $verbose makecache
+    bl64_pkg_run_dnf 'makecache'
     ;;
   ${BL64_OS_CNT}-7.* | ${BL64_OS_OL}-7.*)
-    "$BL64_PKG_CMD_YUM" $verbose makecache
+    bl64_pkg_run_yum 'makecache'
     ;;
   ${BL64_OS_UB}-* | ${BL64_OS_DEB}-*)
-    export DEBIAN_FRONTEND="noninteractive"
-    "$BL64_PKG_CMD_APT" update $verbose
+    bl64_pkg_run_apt 'update'
     ;;
   ${BL64_OS_ALP}-*)
-    "$BL64_PKG_CMD_APK" update $verbose
+    bl64_pkg_run_apk 'update'
     ;;
   ${BL64_OS_MCOS}-*)
-    "$BL64_PKG_CMD_BRW" update $verbose
+    bl64_pkg_run_brew 'update'
     ;;
   *) bl64_check_alert_unsupported ;;
   esac
@@ -91,34 +82,24 @@ function bl64_pkg_prepare() {
 #######################################
 function bl64_pkg_install() {
   bl64_dbg_lib_show_function "$@"
-  local verbose=''
 
-  bl64_check_privilege_root || return $?
-
-  if bl64_dbg_lib_command_enabled; then
-    verbose="$BL64_PKG_SET_VERBOSE"
-  else
-    verbose="$BL64_PKG_SET_QUIET"
-  fi
-
-  bl64_msg_show_task "$_BL64_PKG_TXT_INSTALL"
+  bl64_msg_show_task "$_BL64_PKG_TXT_INSTALL (${*})"
   # shellcheck disable=SC2086
   case "$BL64_OS_DISTRO" in
   ${BL64_OS_FD}-* | ${BL64_OS_RHEL}-* | ${BL64_OS_ALM}-* | ${BL64_OS_RCK}-* | ${BL64_OS_CNT}-8.* | ${BL64_OS_CNT}-9.* | ${BL64_OS_OL}-8.*)
-    "$BL64_PKG_CMD_DNF" $verbose ${BL64_PKG_SET_SLIM} ${BL64_PKG_SET_ASSUME_YES} install -- "$@"
+    bl64_pkg_run_dnf $BL64_PKG_SET_SLIM $BL64_PKG_SET_ASSUME_YES 'install' -- "$@"
     ;;
   ${BL64_OS_CNT}-7.* | ${BL64_OS_OL}-7.*)
-    "$BL64_PKG_CMD_YUM" $verbose ${BL64_PKG_SET_ASSUME_YES} install -- "$@"
+    bl64_pkg_run_yum $BL64_PKG_SET_ASSUME_YES 'install' -- "$@"
     ;;
   ${BL64_OS_UB}-* | ${BL64_OS_DEB}-*)
-    export DEBIAN_FRONTEND="noninteractive"
-    "$BL64_PKG_CMD_APT" install $verbose ${BL64_PKG_SET_ASSUME_YES} -- "$@"
+    bl64_pkg_run_apt 'install' $BL64_PKG_SET_ASSUME_YES -- "$@"
     ;;
   ${BL64_OS_ALP}-*)
-    "$BL64_PKG_CMD_APK" add $verbose -- "$@"
+    bl64_pkg_run_apk 'add' -- "$@"
     ;;
   ${BL64_OS_MCOS}-*)
-    "$BL64_PKG_CMD_BRW" install $verbose "$@"
+    "$BL64_PKG_CMD_BRW" 'install' "$@"
     ;;
   *) bl64_check_alert_unsupported ;;
 
@@ -143,9 +124,54 @@ function bl64_pkg_install() {
 function bl64_pkg_cleanup() {
   bl64_dbg_lib_show_function
   local target=''
+
+  bl64_msg_show_task "$_BL64_PKG_TXT_CLEAN"
+  # shellcheck disable=SC2086
+  case "$BL64_OS_DISTRO" in
+  ${BL64_OS_FD}-* | ${BL64_OS_RHEL}-* | ${BL64_OS_ALM}-* | ${BL64_OS_RCK}-* | ${BL64_OS_CNT}-8.* | ${BL64_OS_CNT}-9.* | ${BL64_OS_OL}-8.*)
+    bl64_pkg_run_dnf 'clean' 'all'
+    ;;
+  ${BL64_OS_CNT}-7.* | ${BL64_OS_OL}-7.*)
+    bl64_pkg_run_yum 'clean' 'all'
+    ;;
+  ${BL64_OS_UB}-* | ${BL64_OS_DEB}-*)
+    bl64_pkg_run_apt 'clean'
+    ;;
+  ${BL64_OS_ALP}-*)
+    bl64_pkg_run_apk 'cache' 'clean'
+    target='/var/cache/apk'
+    if [[ -d "$target" ]]; then
+      bl64_fs_rm_full ${target}/[[:alpha:]]*
+    fi
+    ;;
+  ${BL64_OS_MCOS}-*)
+    bl64_pkg_run_brew 'cleanup' --prune=all -s
+    ;;
+  *) bl64_check_alert_unsupported ;;
+
+  esac
+}
+
+#######################################
+# Command wrapper with verbose, debug and common options
+#
+# * Trust no one. Ignore user provided config and use default config
+#
+# Arguments:
+#   $@: arguments are passed as-is to the command
+# Outputs:
+#   STDOUT: command output
+#   STDERR: command stderr
+# Returns:
+#   command exit status
+#######################################
+function bl64_pkg_run_dnf() {
+  bl64_dbg_lib_show_function "$@"
   local verbose=''
 
-  bl64_check_privilege_root || return $?
+  bl64_check_parameters_none "$#" &&
+    bl64_check_privilege_root ||
+    return $?
 
   if bl64_dbg_lib_command_enabled; then
     verbose="$BL64_PKG_SET_VERBOSE"
@@ -153,30 +179,135 @@ function bl64_pkg_cleanup() {
     verbose="$BL64_PKG_SET_QUIET"
   fi
 
-  bl64_msg_show_task "$_BL64_PKG_TXT_CLEAN"
   # shellcheck disable=SC2086
-  case "$BL64_OS_DISTRO" in
-  ${BL64_OS_FD}-* | ${BL64_OS_RHEL}-* | ${BL64_OS_ALM}-* | ${BL64_OS_RCK}-* | ${BL64_OS_CNT}-8.* | ${BL64_OS_CNT}-9.* | ${BL64_OS_OL}-8.*)
-    "$BL64_PKG_CMD_DNF" clean all $verbose
-    ;;
-  ${BL64_OS_CNT}-7.* | ${BL64_OS_OL}-7.*)
-    "$BL64_PKG_CMD_YUM" clean all $verbose
-    ;;
-  ${BL64_OS_UB}-* | ${BL64_OS_DEB}-*)
-    export DEBIAN_FRONTEND="noninteractive"
-    "$BL64_PKG_CMD_APT" clean
-    ;;
-  ${BL64_OS_ALP}-*)
-    "$BL64_PKG_CMD_APK" cache clean $verbose
-    target='/var/cache/apk'
-    if [[ -d "$target" ]]; then
-      bl64_fs_rm_full ${target}/[[:alpha:]]*
-    fi
-    ;;
-  ${BL64_OS_MCOS}-*)
-    "$BL64_PKG_CMD_BRW" cleanup $verbose --prune=all -s
-    ;;
-  *) bl64_check_alert_unsupported ;;
+  "$BL64_PKG_CMD_DNF" $verbose "$@"
+}
 
-  esac
+#######################################
+# Command wrapper with verbose, debug and common options
+#
+# * Trust no one. Ignore user provided config and use default config
+#
+# Arguments:
+#   $@: arguments are passed as-is to the command
+# Outputs:
+#   STDOUT: command output
+#   STDERR: command stderr
+# Returns:
+#   command exit status
+#######################################
+function bl64_pkg_run_yum() {
+  bl64_dbg_lib_show_function "$@"
+  local verbose=''
+
+  bl64_check_parameters_none "$#" &&
+    bl64_check_privilege_root ||
+    return $?
+
+  if bl64_dbg_lib_command_enabled; then
+    verbose="$BL64_PKG_SET_VERBOSE"
+  else
+    verbose="$BL64_PKG_SET_QUIET"
+  fi
+
+  # shellcheck disable=SC2086
+  "$BL64_PKG_CMD_YUM" $verbose "$@"
+}
+
+#######################################
+# Command wrapper with verbose, debug and common options
+#
+# * Trust no one. Ignore user provided config and use default config
+#
+# Arguments:
+#   $@: arguments are passed as-is to the command
+# Outputs:
+#   STDOUT: command output
+#   STDERR: command stderr
+# Returns:
+#   command exit status
+#######################################
+function bl64_pkg_run_apt() {
+  bl64_dbg_lib_show_function "$@"
+  local verbose=''
+
+  bl64_check_parameters_none "$#" &&
+    bl64_check_privilege_root ||
+    return $?
+
+  if bl64_dbg_lib_command_enabled; then
+    verbose="$BL64_PKG_SET_VERBOSE"
+  else
+    DEBCONF_NOWARNINGS='yes'
+    DEBCONF_TERSE='yes'
+    verbose="$BL64_PKG_SET_QUIET"
+  fi
+
+  # Avoid interactive questions
+  DEBIAN_FRONTEND="noninteractive"
+
+  # shellcheck disable=SC2086
+  "$BL64_PKG_CMD_APT" $verbose "$@"
+}
+
+#######################################
+# Command wrapper with verbose, debug and common options
+#
+# * Trust no one. Ignore user provided config and use default config
+#
+# Arguments:
+#   $@: arguments are passed as-is to the command
+# Outputs:
+#   STDOUT: command output
+#   STDERR: command stderr
+# Returns:
+#   command exit status
+#######################################
+function bl64_pkg_run_apk() {
+  bl64_dbg_lib_show_function "$@"
+  local verbose=''
+
+  bl64_check_parameters_none "$#" &&
+    bl64_check_privilege_root ||
+    return $?
+
+  if bl64_dbg_lib_command_enabled; then
+    verbose="$BL64_PKG_SET_VERBOSE"
+  else
+    verbose="$BL64_PKG_SET_QUIET"
+  fi
+
+  # shellcheck disable=SC2086
+  "$BL64_PKG_CMD_APK" $verbose "$@"
+}
+
+#######################################
+# Command wrapper with verbose, debug and common options
+#
+# * Trust no one. Ignore user provided config and use default config
+#
+# Arguments:
+#   $@: arguments are passed as-is to the command
+# Outputs:
+#   STDOUT: command output
+#   STDERR: command stderr
+# Returns:
+#   command exit status
+#######################################
+function bl64_pkg_run_brew() {
+  bl64_dbg_lib_show_function "$@"
+  local verbose=''
+
+  bl64_check_parameters_none "$#" &&
+    bl64_check_privilege_root ||
+    return $?
+
+  if bl64_dbg_lib_command_enabled; then
+    verbose="$BL64_PKG_SET_VERBOSE"
+  else
+    verbose="$BL64_PKG_SET_QUIET"
+  fi
+
+  # shellcheck disable=SC2086
+  "$BL64_PKG_CMD_BRW" $verbose "$@"
 }